sql: fix schema and simplify tests Signed-off-by: Brendan Cronin <brendan.cronin@mercedes-benz.com>
