+++ extensions/sql/contract-definition/loader/src/test/java/org/eclipse/dataspaceconnector/sql/contractdefinition/loader/SqlContractDefinitionLoaderTest.java
@@ -14,8 +14,14 @@
 
 package org.eclipse.dataspaceconnector.sql.contractdefinition.loader;
 
+import org.eclipse.dataspaceconnector.dataloading.ContractDefinitionLoader;
+import org.eclipse.dataspaceconnector.junit.launcher.EdcExtension;
 import org.eclipse.dataspaceconnector.policy.model.Policy;
 import org.eclipse.dataspaceconnector.spi.asset.AssetSelectorExpression;
+import org.eclipse.dataspaceconnector.spi.system.ServiceExtension;
+import org.eclipse.dataspaceconnector.spi.system.ServiceExtensionContext;
+import org.eclipse.dataspaceconnector.spi.transaction.TransactionContext;
+import org.eclipse.dataspaceconnector.spi.transaction.datasource.DataSourceRegistry;
 import org.eclipse.dataspaceconnector.spi.types.domain.contract.offer.ContractDefinition;
 import org.eclipse.dataspaceconnector.sql.contractdefinition.schema.SqlContractDefinitionTables;
 import org.junit.jupiter.api.AfterEach;
@@ -23,15 +29,18 @@ import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 
 import java.sql.Connection;
 import java.sql.SQLException;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.concurrent.atomic.AtomicReference;
 
 import static org.eclipse.dataspaceconnector.sql.SqlQueryExecutor.executeQuery;
 
-class SqlContractDefinitionLoaderTest extends AbstractSqlContractDefinitionLoaderTest {
+@ExtendWith(EdcExtension.class)
+class SqlContractDefinitionLoaderTest {
 
     private static final String DATASOURCE_NAME = "contractdefinition";
 
@@ -43,9 +52,16 @@ class SqlContractDefinitionLoaderTest extends AbstractSqlContractDefinitionLoade
         }
     };
 
+    private final AtomicReference<ServiceExtensionContext> contextRef = new AtomicReference<>();
+
     @BeforeEach
-    void setUp() {
+    void setUp(EdcExtension extension) {
         systemProperties.forEach(System::setProperty);
+        extension.registerSystemExtension(ServiceExtension.class, new ServiceExtension() {
+            public void initialize(ServiceExtensionContext context) {
+                contextRef.set(context);
+            }
+        });
     }
 
     @AfterEach
@@ -58,6 +74,19 @@ class SqlContractDefinitionLoaderTest extends AbstractSqlContractDefinitionLoade
             }
         });
         systemProperties.keySet().forEach(System::clearProperty);
+        contextRef.set(null);
+    }
+
+    protected DataSourceRegistry getDataSourceRegistry() {
+        return contextRef.get().getService(DataSourceRegistry.class);
+    }
+
+    protected TransactionContext getTransactionContext() {
+        return contextRef.get().getService(TransactionContext.class);
+    }
+
+    protected ContractDefinitionLoader getContractDefinitionLoader() {
+        return contextRef.get().getService(ContractDefinitionLoader.class);
     }
 
     @Test
